add dead

+++ Birgit/Birgit/SpriteSheet.cpp
@@ -11,8 +11,11 @@ SpriteSheet::SpriteSheet(const sf::Texture& texture, size_t tileWidth, size_t ti
     , m_spriteCol(0)
     , m_spriteRow(0)
 {
-    m_sprite.setTextureRect(sf::IntRect(m_spriteCol * m_tileWidth, m_spriteRow * m_tileHeight, m_tileWidth, m_tileHeight));
+    m_sprite.setTextureRect(sf::IntRect(m_spriteCol * static_cast<int>(m_tileWidth), 
+        m_spriteRow * static_cast<int>(m_tileHeight), static_cast<int>(m_tileWidth), static_cast<int>(m_tileHeight)));
     m_sprite.setScale(0.2f, 0.2f);  // 20% size
+    // TODO experiment with the below
+    // TODO NEXT make all sprites much smaller!!!
     m_sprite.setOrigin(tileWidth / 2.f, tileHeight / 2.f);
 }
 
@@ -20,7 +23,7 @@ void SpriteSheet::reset(const sf::Vector2f pos)
 {
     if (m_playReverse)
     {
-        m_spriteCol = m_spriteCount - 1;
+        m_spriteCol = static_cast<int>(m_spriteCount) - 1;
         m_spriteRow = 0;
     }
     else
@@ -43,7 +46,7 @@ void SpriteSheet::update(float dt)
         if (m_playReverse)
         {
             if (m_spriteCol == 0)
-                m_spriteCol = m_spriteCount - 1;
+                m_spriteCol = static_cast<int>(m_spriteCount) - 1;
             else
                 --m_spriteCol;
         }
@@ -56,10 +59,10 @@ void SpriteSheet::update(float dt)
     }
 
     m_sprite.setTextureRect(sf::IntRect(
-        m_spriteCol * m_tileWidth,
-        m_spriteRow * m_tileHeight,
-        m_tileWidth,
-        m_tileHeight
+        m_spriteCol * static_cast<int>(m_tileWidth),
+        m_spriteRow * static_cast<int>(m_tileHeight),
+        static_cast<int>(m_tileWidth),
+        static_cast<int>(m_tileHeight)
     ));
 }
